Bug Fix: Monkey Respawn System
Fixed (or at least think I've fixed. It'll probably find a way to prove me wrong) the bug with the Monkey/Alien spawning atop the player at the checkpoints. Will now search in the code for the player's respawn point and won't allow the respawn point for the monkey to be set until it's different from the player's.